+++ src/directives/transition.js
@@ -1,9 +1,12 @@
 module.exports = {
 
   priority: 1000,
+  isLiteral: true,
 
   bind: function () {
-    this.el.__vueTransition = this.expression
+    this.el.__v_trans = {
+      id: this.expression
+    }
   }
 
 }